Move S3Uploader into the S3 namespace
