+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/storage/LargeResponsesTest.scala
@@ -17,8 +17,8 @@ import akka.stream.scaladsl.StreamConverters
 import org.apache.commons.io.IOUtils
 import org.scalatest.funspec.AnyFunSpec
 import uk.ac.wellcome.akka.fixtures.Akka
-import uk.ac.wellcome.fixtures.{TestWith}
-import uk.ac.wellcome.platform.archive.common.storage.services.S3Uploader
+import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.platform.archive.common.storage.services.s3.S3Uploader
 import uk.ac.wellcome.storage.ObjectLocationPrefix
 import uk.ac.wellcome.storage.fixtures.S3Fixtures
 import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
